+++ src/async_impl/client.rs
@@ -105,6 +105,7 @@ struct Config {
     cookie_store: Option<cookie::CookieStore>,
     trust_dns: bool,
     error: Option<crate::Error>,
+    https_only: bool,
 }
 
 impl Default for ClientBuilder {
@@ -134,7 +135,9 @@ impl ClientBuilder {
                 connection_verbose: false,
                 pool_idle_timeout: Some(Duration::from_secs(90)),
                 pool_max_idle_per_host: std::usize::MAX,
-                tcp_keepalive: Some(Duration::from_secs(60)),
+                // TODO: Re-enable default duration once hyper's HttpConnector is fixed
+                // to no longer error when an option fails.
+                tcp_keepalive: None, //Some(Duration::from_secs(60)),
                 proxies: Vec::new(),
                 auto_sys_proxy: true,
                 redirect_policy: redirect::Policy::default(),
@@ -155,6 +158,7 @@ impl ClientBuilder {
                 trust_dns: cfg!(feature = "trust-dns"),
                 #[cfg(feature = "cookies")]
                 cookie_store: None,
+                https_only: false,
             },
         }
     }
@@ -343,6 +347,7 @@ impl ClientBuilder {
                 request_timeout: config.timeout,
                 proxies,
                 proxies_maybe_http_auth,
+                https_only: config.https_only,
             }),
         })
     }
@@ -568,12 +573,6 @@ impl ClientBuilder {
         self
     }
 
-    #[doc(hidden)]
-    #[deprecated(note = "the system proxy is used automatically")]
-    pub fn use_sys_proxy(self) -> ClientBuilder {
-        self
-    }
-
     // Timeout options
 
     /// Enables a request timeout.
@@ -632,12 +631,6 @@ impl ClientBuilder {
         self
     }
 
-    #[doc(hidden)]
-    #[deprecated(note = "renamed to `pool_max_idle_per_host`")]
-    pub fn max_idle_per_host(self, max: usize) -> ClientBuilder {
-        self.pool_max_idle_per_host(max)
-    }
-
     /// Enable case sensitive headers.
     pub fn http1_title_case_headers(mut self) -> ClientBuilder {
         self.config.http1_title_case_headers = true;
@@ -671,22 +664,10 @@ impl ClientBuilder {
 
     // TCP options
 
-    #[doc(hidden)]
-    #[deprecated(note = "tcp_nodelay is enabled by default, use `tcp_nodelay_` to disable")]
-    pub fn tcp_nodelay(self) -> ClientBuilder {
-        self.tcp_nodelay_(true)
-    }
-
     /// Set whether sockets have `SO_NODELAY` enabled.
     ///
     /// Default is `true`.
-    // NOTE: Regarding naming (trailing underscore):
-    //
-    // Due to the original `tcp_nodelay()` not taking an argument, changing
-    // the default means a user has no way of *disabling* this feature.
-    //
-    // TODO(v0.11.x): Remove trailing underscore.
-    pub fn tcp_nodelay_(mut self, enabled: bool) -> ClientBuilder {
+    pub fn tcp_nodelay(mut self, enabled: bool) -> ClientBuilder {
         self.config.nodelay = enabled;
         self
     }
@@ -713,8 +694,6 @@ impl ClientBuilder {
     /// Set that all sockets have `SO_KEEPALIVE` set with the supplied duration.
     ///
     /// If `None`, the option will not be set.
-    ///
-    /// Default is 60 seconds.
     pub fn tcp_keepalive<D>(mut self, val: D) -> ClientBuilder
         where
             D: Into<Option<Duration>>,
@@ -903,6 +882,14 @@ impl ClientBuilder {
             self
         }
     }
+
+    /// Restrict the Client to be used with HTTPS only requests.
+    /// 
+    /// Defaults to false.
+    pub fn https_only(mut self, enabled: bool) -> ClientBuilder {
+        self.config.https_only = enabled;
+        self
+    }
 }
 
 type HyperClient = hyper::Client<Connector, super::body::ImplStream>;
@@ -1026,6 +1013,11 @@ impl Client {
             return Pending::new_err(error::url_bad_scheme(url));
         }
 
+        // check if we're in https_only mode and check the scheme of the current URL
+        if self.inner.https_only && url.scheme() != "https" {
+            return Pending::new_err(error::url_bad_scheme(url));
+        }
+
         // insert default headers in the request headers
         // without overwriting already appended headers.
         for (key, value) in &self.inner.headers {
@@ -1073,7 +1065,8 @@ impl Client {
 
         let timeout = timeout
             .or(self.inner.request_timeout)
-            .map(tokio::time::sleep);
+            .map(tokio::time::sleep)
+            .map(Box::pin);
 
         *req.headers_mut() = headers.clone();
 
@@ -1224,6 +1217,7 @@ struct ClientRef {
     request_timeout: Option<Duration>,
     proxies: Arc<Vec<Proxy>>,
     proxies_maybe_http_auth: bool,
+    https_only: bool,
 }
 
 impl ClientRef {
@@ -1286,7 +1280,7 @@ pin_project! {
         #[pin]
         in_flight: ResponseFuture,
         #[pin]
-        timeout: Option<Sleep>,
+        timeout: Option<Pin<Box<Sleep>>>,
     }
 }
 
@@ -1295,7 +1289,7 @@ impl PendingRequest {
         self.project().in_flight
     }
 
-    fn timeout(self: Pin<&mut Self>) -> Pin<&mut Option<Sleep>> {
+    fn timeout(self: Pin<&mut Self>) -> Pin<&mut Option<Pin<Box<Sleep>>>> {
         self.project().timeout
     }